+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-18 20:57:51
-// Commit: fix(state): fix memory leak by cancelling state subscription
+// Updated at: 2026-08-18 21:33:23
+// Commit: refactor(state): optimize data flow between domain and data layers
 
 class stateComponentState {
-  final String commitRef = '3720';
-  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
-  final DateTime timestamp = DateTime.parse('2026-08-18T20:57:51+05:00');
+  final String commitRef = '3724';
+  final String lastMessage = 'refactor(state): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-08-18T21:33:23+05:00');
 }